Move throws-on-prototype up one level
