+++ lib/prototypes/copy-prototype-methods.mjs
@@ -1,4 +1,4 @@
-import throwsOnProto from "./throws-on-proto.mjs";
+import throwsOnProto from "../throws-on-proto.mjs";
 var call = Function.call;
 
 var disallowedProperties = [
