+++ lib/prototypes/index.test.mjs
@@ -1,7 +1,7 @@
 import pkg from "@sinonjs/referee-sinon";
 const { assert } = pkg;
 
-import throwsOnProto from "./throws-on-proto.mjs";
+import throwsOnProto from "../throws-on-proto.mjs";
 
 import prototypes from "./index.mjs";
 var arrayProto = prototypes.array;
