+++ lib/throws-on-proto.mjs
